+++ gapipy/resources/__init__.py
@@ -1,7 +1,9 @@
 from .tour import (  # NOQA
     Accommodation,
     Activity,
+    OptionalActivity,
     Departure,
+    Duration,
     Itinerary,
     Promotion,
     SingleSupplement,
@@ -16,6 +18,8 @@ from .dossier import (  # NOQA
     ActivityDossier,
     PlaceDossier,
     TransportDossier,
+    DossierDetail,
+    DossierDetailType,
 )
 
 from .booking import (  # NOQA